chore: Reorganize workspace packages to avoid having logic in `__init__.py` files
